Separate mcnp2cad to its own file
